+++ qbraid/transpiler/conversions/openqasm3/openqasm3_to_ionq.py
@@ -35,6 +35,7 @@ IONQ_ONE_QUBIT_GATE_MAP = {
     "not": "not",
     "y": "y",
     "z": "z",
+    "p": "z",
     "h": "h",
     "s": "s",
     "si": "si",
@@ -56,9 +57,9 @@ IONQ_ONE_QUBIT_GATE_MAP = {
 IONQ_TWO_QUBIT_GATE_MAP = {
     "cnot": "cnot",
     "cx": "cnot",
-    "CX": "cnot",
     "swap": "swap",
     "zz": "zz",
+    "rzz": "zz",
     "ms": "ms",
 }
 
@@ -124,7 +125,7 @@ def _parse_gates(program: Union[OpenQasm2Program, OpenQasm3Program]) -> list[dic
 
     for statement in original_program.statements:
         if isinstance(statement, openqasm3.ast.QuantumGate):
-            name = statement.name.name
+            name = statement.name.name.lower()
             qubits = statement.qubits
             qubit_values = []